removal of even more duplicate code by centralizing in utils
