+++ dataland-frontend/src/components/pages/MyDataRequestsOverview.vue
@@ -181,12 +181,7 @@ import DataTable, {
   type DataTableSortEvent,
 } from 'primevue/datatable';
 import Column from 'primevue/column';
-import {
-  frameworkHasSubTitle,
-  getFrameworkSubtitle,
-  getFrameworkTitle,
-  humanizeStringOrNumber,
-} from '@/utils/StringFormatter';
+import { frameworkHasSubTitle, getFrameworkSubtitle, getFrameworkTitle } from '@/utils/StringFormatter';
 import DatasetsTabMenu from '@/components/general/DatasetsTabMenu.vue';
 import { convertUnixTimeInMsToDateString } from '@/utils/DataFormatUtils';
 import { AccessStatus, type ExtendedStoredDataRequest, RequestStatus } from '@clients/communitymanager';
@@ -194,20 +189,19 @@ import { type DataTypeEnum } from '@clients/backend';
 import InputText from 'primevue/inputtext';
 import FrameworkDataSearchDropdownFilter from '@/components/resources/frameworkDataSearch/FrameworkDataSearchDropdownFilter.vue';
 import { type FrameworkSelectableItem, type SelectableItem } from '@/utils/FrameworkDataSearchDropDownFilterTypes';
-import { FRAMEWORKS_WITH_VIEW_PAGE } from '@/utils/Constants';
-import { getFrontendFrameworkDefinition } from '@/frameworks/FrontendFrameworkRegistry';
 import AuthenticationWrapper from '@/components/wrapper/AuthenticationWrapper.vue';
 import { accessStatusBadgeClass, badgeClass } from '@/utils/RequestUtils';
-import { customCompareForRequestStatus } from '@/utils/RequestsOverviewPageUtils';
+import {
+  customCompareForRequestStatus,
+  retrieveAvailableAccessStatus,
+  retrieveAvailableFrameworks,
+} from '@/utils/RequestsOverviewPageUtils';
 
 export default defineComponent({
   name: 'MyDataRequestsOverview',
   computed: {
     RequestStatus() {
       return RequestStatus;
-    },
-    AccessStatus() {
-      return AccessStatus;
     },
   },
   components: {
@@ -252,8 +246,8 @@ export default defineComponent({
     };
   },
   mounted() {
-    this.availableFrameworks = this.retrieveAvailableFrameworks();
-    this.availableAccessStatus = this.retrieveAvailableAccessStatus();
+    this.availableFrameworks = retrieveAvailableFrameworks();
+    this.availableAccessStatus = retrieveAvailableAccessStatus();
     this.getStoredRequestDataList().catch((error) => console.error(error));
     this.resetFilterAndSearchBar();
   },
@@ -298,36 +292,6 @@ export default defineComponent({
       return this.$router.push(url);
     },
 
-    /**
-     * Gets list with all available frameworks
-     * @returns array of frameworkSelectableItem
-     */
-    retrieveAvailableFrameworks(): Array<FrameworkSelectableItem> {
-      return FRAMEWORKS_WITH_VIEW_PAGE.map((dataTypeEnum) => {
-        let displayName = humanizeStringOrNumber(dataTypeEnum);
-        const frameworkDefinition = getFrontendFrameworkDefinition(dataTypeEnum);
-        if (frameworkDefinition) {
-          displayName = frameworkDefinition.label;
-        }
-        return {
-          frameworkDataType: dataTypeEnum,
-          displayName: displayName,
-          disabled: false,
-        };
-      });
-    },
-    /**
-     * Gets list with all available access status
-     * @returns array of SelectableItem
-     */
-    retrieveAvailableAccessStatus(): Array<SelectableItem> {
-      return Object.values(this.AccessStatus).map((status) => {
-        return {
-          displayName: status,
-          disabled: false,
-        };
-      });
-    },
     /**
      * Gets list of storedDataRequests
      */
